fixed one-hot issue?

+++ experiments/experiment_code/generic_scan.py
@@ -27,14 +27,14 @@ from sv3.nn import FunctionalModelJac
 SVD_LOSS_FNS = {
     "ce": lambda pred, y: F.cross_entropy(pred, y, reduction='none'),
     "mse": lambda pred, y: ((pred - y) ** 2).sum(dim=-1),
-    "label_regression_10class": lambda pred, y: (pred - F.one_hot(y,num_classes=10).to(pred)).pow(2).sum(dim=1)
+    "label_regression_10class": lambda pred, y: (pred - F.one_hot(y.to(torch.long),num_classes=10).to(pred)).pow(2).sum(dim=1)
 }
 
 # Standard loss returns a scalar
 STANDARD_LOSS_FNS = {
     "ce": lambda: nn.CrossEntropyLoss(),
     "mse": lambda: nn.MSELoss(),
-    "label_regression_10class": lambda pred, y: (pred - F.one_hot(y,num_classes=10).to(pred)).pow(2).sum(dim=1).mean()
+    "label_regression_10class": lambda pred, y: (pred - F.one_hot(y.to(torch.long),num_classes=10).to(pred)).pow(2).sum(dim=1).mean()
 }
